Refactor project & examples. Add documentation
#37
