--- GofConsoleApp/Examples/Structural/ProxyPattern/ConfigProviderCachedOutput/EnumConfigEnv.cs
@@ -1,9 +1,0 @@
-namespace GofConsoleApp.Examples.Structural.ProxyPattern.ConfigProviderCachedOutput;
-
-internal enum EnumConfigEnv
-{
-    Invalid,
-    Dev,
-    Prod,
-    Test
-}
