+++ GofConsoleApp/Examples/Structural/ProxyPattern/ConfigProviderComponents/EnumConfigEnv.cs
@@ -1,0 +1,9 @@
+namespace GofConsoleApp.Examples.Structural.ProxyPattern.ConfigProviderComponents;
+
+internal enum EnumConfigEnv
+{
+    Invalid,
+    Dev,
+    Prod,
+    Test
+}
