--- GofConsoleApp/Examples/Structural/ProxyPattern/ExampleUserInterface/EnumUserType.cs
@@ -1,9 +1,0 @@
-﻿namespace GofConsoleApp.Examples.Structural.ProxyPattern.ExampleUserInterface;
-
-internal enum EnumUserType
-{
-    Invalid,
-    Admin,
-    Standard,
-    Guest
-}
