--- GofConsoleApp/Examples/Structural/ProxyPattern/ExampleUserInterface/IUserInterface.cs
@@ -1,5 +1,0 @@
-﻿using GofPatterns.Structural.ProxyPattern.Interfaces;
-
-namespace GofConsoleApp.Examples.Structural.ProxyPattern.ExampleUserInterface;
-
-internal interface IUserInterface : IProxyComponent<EnumOperationOption, string> { }
