+++ GofConsoleApp/Examples/Structural/ProxyPattern/NewsChannelComponents/INewsChannelProxy.cs
@@ -1,5 +1,5 @@
 ﻿using GofPatterns.Structural.ProxyPattern.Interfaces;
 
-namespace GofConsoleApp.Examples.Structural.ProxyPattern.NewsChannel;
+namespace GofConsoleApp.Examples.Structural.ProxyPattern.NewsChannelComponents;
 
 internal interface INewsChannelProxy : INewsChannel, IProxyBoundedAccess<EnumNewsChannel> { }
