+++ GofConsoleApp/Examples/Structural/ProxyPattern/UserInterfaceComponents/EnumUserType.cs
@@ -1,0 +1,9 @@
+﻿namespace GofConsoleApp.Examples.Structural.ProxyPattern.UserInterfaceComponents;
+
+internal enum EnumUserType
+{
+    Invalid,
+    Admin,
+    Standard,
+    Guest
+}
