+++ GofConsoleApp/Examples/Structural/ProxyPattern/UserInterfaceComponents/IUserInterfaceProxy.cs
@@ -1,0 +1,5 @@
+﻿using GofPatterns.Structural.ProxyPattern.Interfaces;
+
+namespace GofConsoleApp.Examples.Structural.ProxyPattern.UserInterfaceComponents;
+
+internal interface IUserInterfaceProxy : IUserInterface, IProxyBoundedAccess<EnumOperationOption, string> { }
