+++ GofConsoleApp/Examples/Structural/ProxyPattern/UserInterfaceComponents/UserInterfaceProxyGuest.cs
@@ -1,0 +1,9 @@
+﻿using GofPatterns.Structural.ProxyPattern;
+using static GofConsoleApp.Examples.Structural.ProxyPattern.UserInterfaceComponents.EnumOperationOption;
+
+namespace GofConsoleApp.Examples.Structural.ProxyPattern.UserInterfaceComponents;
+
+internal class UserInterfaceProxyGuest : ProxyBoundedAccess<EnumOperationOption, string>, IUserInterfaceProxy
+{
+    public UserInterfaceProxyGuest(IUserInterface userInterface) : base(userInterface, new[] { Read }) { }
+}
